Simplify fetching function
