+++ lib/api.js
@@ -1,5 +1,9 @@
+import { keyBy } from 'lodash';
+import bbox from '@turf/bbox';
+
 import { airtableGet } from '#/lib/airtable.js';
 import { vtaGet } from '#/lib/vta.js';
+import { isValidBbox, preprocessData } from '#/lib/util.js';
 
 export function fetchAllocations() {
   return airtableGet('Allocations', {
@@ -111,4 +115,99 @@ export function fetchFaq() {
     console.error(error);
     return [];
   });
+}
+
+export function fetchGeoJson(items) {
+  return Promise.all(
+    items.map(async (item) => {
+      try {
+        if (item.fields.geojson) {
+          const response = await fetch(item.fields.geojson[0].url, {
+            next: { revalidate: 600 },
+          });
+
+          if (response.ok) {
+            const geojson = await response.json();
+            const layerBbox = bbox(geojson);
+
+            if (!isValidBbox(layerBbox)) {
+              throw new Error('geojson outside of bbox');
+            }
+
+            return {
+              id: item.id,
+              bbox: layerBbox,
+              geojson,
+            };
+          }
+        } else if (item.fields.Latitude && item.fields.Longitude) {
+          const geojson = {
+            type: 'Feature',
+            properties: {
+              projectId: item.id,
+            },
+            geometry: {
+              type: 'Point',
+              coordinates: [item.fields.Longitude, item.fields.Latitude],
+            },
+          };
+
+          return {
+            id: item.id,
+            bbox: bbox(geojson),
+            geojson,
+          };
+        }
+      } catch (error) {
+        console.warn(`Invalid geometry for item "${item.fields.Name}"`);
+        console.warn(item.fields.geojson);
+        console.warn(error);
+      }
+    })
+  );
+}
+
+export async function fetchData() {
+  const [
+    allocations,
+    awards,
+    categories,
+    documents,
+    grantees,
+    expenditures,
+    projects,
+    revenue,
+    faqs,
+  ] = await Promise.all([
+    fetchAllocations(),
+    fetchAwards(),
+    fetchCategories(),
+    fetchDocuments(),
+    fetchGrantees(),
+    fetchExpenditures(),
+    fetchProjects(),
+    fetchRevenue(),
+    fetchFaq(),
+  ]);
+
+  const data = preprocessData({
+    allocations,
+    awards,
+    categories,
+    documents,
+    grantees,
+    expenditures,
+    projects,
+    revenue,
+    faqs,
+  });
+
+  const mapData = [
+    ...(await fetchGeoJson(data.projects)),
+    ...(await fetchGeoJson(data.grantees)),
+  ];
+
+  data.geojsons = keyBy(mapData, 'id');
+
+  return data;
 }
