+++ lib/util.js
@@ -1,9 +1,8 @@
-/* global fetch, window */
+/* global window */
 import querystring from 'querystring';
 import algoliasearch from 'algoliasearch';
 
 import { maxBy, sortBy, uniq } from 'lodash';
-import bbox from '@turf/bbox';
 import moment from 'moment';
 
 import { categoryCards } from '#/lib/category-cards.js';
@@ -487,56 +486,6 @@ export function preprocessData(data) {
   });
 
   return data;
-}
-
-export async function fetchGeoJson(items) {
-  return Promise.all(
-    items.map(async (item) => {
-      try {
-        if (item.fields.geojson) {
-          const response = await fetch(item.fields.geojson[0].url, {
-            next: { revalidate: 600 },
-          });
-
-          if (response.ok) {
-            const geojson = await response.json();
-            const layerBbox = bbox(geojson);
-
-            if (!isValidBbox(layerBbox)) {
-              throw new Error('geojson outside of bbox');
-            }
-
-            return {
-              id: item.id,
-              bbox: layerBbox,
-              geojson,
-            };
-          }
-        } else if (item.fields.Latitude && item.fields.Longitude) {
-          const geojson = {
-            type: 'Feature',
-            properties: {
-              projectId: item.id,
-            },
-            geometry: {
-              type: 'Point',
-              coordinates: [item.fields.Longitude, item.fields.Latitude],
-            },
-          };
-
-          return {
-            id: item.id,
-            bbox: bbox(geojson),
-            geojson,
-          };
-        }
-      } catch (error) {
-        console.warn(`Invalid geometry for item "${item.fields.Name}"`);
-        console.warn(item.fields.geojson);
-        console.warn(error);
-      }
-    })
-  );
 }
 
 export function isValidBbox(bbox) {
